Well, it has been a while...
Implemented OnChangeArea into MapManagerScript.
Added 0'd out calls to OnChangeArea for scripts and MapManagerScripts when removing player from world.
Removed Wintergrasp direct implementation and implemented it through the external script system for better handling.
Fixed a few things in regards to Lacrimi scripts sleeping the core instead of a threadsafe delay, and configs not being loaded directly to allow for more accurate checks in load scripts.

+++ src/Lacrimi/Source/Scripts/SpellScripts/Paladin.cpp
@@ -335,8 +335,8 @@ void EyeForAnEye(uint32 i, Spell* pSpell, uint32 effect)
 	{
 		if( pSpell->p_caster != NULL)
 		{
-			if(pSpell->damage > (pSpell->p_caster->GetUInt32Value(UNIT_FIELD_MAXHEALTH)/2))
-				pSpell->damage = (pSpell->p_caster->GetUInt32Value(UNIT_FIELD_MAXHEALTH)/2);
+			if((uint32)pSpell->damage > (pSpell->p_caster->GetUInt32Value(UNIT_FIELD_MAXHEALTH)/2))
+				pSpell->damage = int32(pSpell->p_caster->GetUInt32Value(UNIT_FIELD_MAXHEALTH)/2);
 		}
 	}
 }
